Move the function to performance_lib.py
